refactoring: refactored files and folders structure according to FSD
