+++ src/app/layout.tsx
@@ -9,7 +9,7 @@ import { Geist, Geist_Mono } from 'next/font/google';
 // import "@/shared/styles/globals.css";
 import '@mantine/core/styles.css';
 import '@/shared/styles/reset.css';
-import type { BaseLayoutProps } from '@/shared/types/core';
+import { PropsWithChildren } from 'react';
 
 const geistSans = Geist({
     variable: '--font-geist-sans',
@@ -26,7 +26,7 @@ export const metadata: Metadata = {
     description: 'Вело майстерня VeloMax',
 };
 
-export default function RootLayout({ children }: BaseLayoutProps) {
+export default function RootLayout({ children }: PropsWithChildren) {
     return (
         <html lang="en" {...mantineHtmlProps}>
             <head>
