+++ src/entities/admin/ admin-main-layout/config/nav.ts
@@ -1,0 +1,13 @@
+import { URLs } from "@/shared/constants/urls";
+
+export interface NavItem {
+    label: string;
+    link: string;
+}
+
+export const navItems: NavItem[] = [
+    { label: 'Dashboard', link: URLs.admin.main },
+    { label: 'Products', link: URLs.admin.products },
+    { label: 'Categories', link: URLs.admin.categories },
+    { label: 'Orders', link: URLs.admin.orders },
+];
