+++ src/features/admin-panel/categories/index.ts
@@ -1,0 +1,2 @@
+export { default as Categories } from './Categories';
+export { getCategories } from './services/categories.service';
