+++ src/features/admin-panel/products/index.ts
@@ -1,0 +1,2 @@
+export { default as Products } from './Products';
+export { getProducts } from './services/products.service';
