+++ src/shared/types/category.ts
@@ -1,0 +1,5 @@
+export type Category = {
+    id: number;
+    name: string;
+    description: string;
+};
