--- src/shared/types/core.ts
@@ -1,3 +1,0 @@
-export type BaseLayoutProps = Readonly<{
-    children: React.ReactNode;
-}>;
