+++ src/shared/types/product.ts
@@ -1,0 +1,7 @@
+export type Product = {
+    id: number;
+    name: string;
+    brand: string;
+    price: number;
+    category: string;
+};
